Move server functions into separate module.
